:lipstick: Style: 반응형 디자인 수정
