+++ daeng/src/components/review/WriteReview.jsx
@@ -4,7 +4,10 @@ import PreferenceFavoriteOptionList from "./PreferenceFavoriteOptionList";
 import star from "../../assets/icons/star.svg";
 import notfillstar from "../../assets/icons/notfillstar.svg";
 import addImg from "../../assets/icons/addImg.svg";
-import TextContainer from "./TextContainer";
+import axios from "axios";
+import ConfirmBtn from '../../components/commons/ConfirmBtn';
+import AlertDialog from '../../components/commons/SweetAlert';
+
 
 const WriteReviewAllContainer = styled.div`
   display: block;
@@ -25,14 +28,16 @@ const PlaceTitle = styled.span`
   margin-left: 10px;
 
   @media (max-width: 554px) {
-    font-size: 15px;
-    margin-right: 40%;
+    font-size: 20px;
+    margin-right: 50%;
+    margin-left:10px;
   }
 `;
 
 const WriteReviewDate = styled.span`
   color: #b3b3b3;
   font-size: 13px;
+  margin-left:20px;
 `;
 
 const SelectPlaceOptionContainer = styled.div`
@@ -42,6 +47,11 @@ const SelectPlaceOptionContainer = styled.div`
   text-align: left;
   padding: 5%;
   margin-right: 10px;
+
+  @media (max-width: 554px) {
+    width: auto;
+    height: 230px;
+  }
 `;
 
 const WhatPointLike = styled.span`
@@ -130,8 +140,8 @@ const Question = styled.span`
 const PetSelection = styled.select`
   width: 40%;
   height: 30px;
-  padding: 10px;
   font-size: 15px;
+  padding-left: 10px;
   border: 0.5px solid #d9d9d9;
   border-radius: 5px;
   cursor: pointer;
@@ -204,18 +214,81 @@ const AddImgButton = styled.img`
   margin-bottom: 0px;
   cursor: pointer;
 `;
+//text container css
+const QuestionBox = styled.span`
+  font-size: 15px;
+  display: inline; 
+  color: #333;
+
+  p {
+    display: inline-block;
+    font-size: 13px;
+    color: #D9D9D9;
+    margin-left: 5px;
+  }
+`
+
+const CountText = styled.span`
+  font-size: 11px;
+  color: #FF0000;
+  margin-top:3px;
+  margin-right:10px;
+`
+
+const TextDescriptionContainer = styled.div`
+  margin-top: 20px;
+  display: flex;
+  margin-bottom: -12px;
+  justify-content: space-between;  
+  align-items: center;  
+`
+const DivisionLine = styled.div`
+  height: 1px;
+  background-color: #E5E5E5;
+  margin-top:20px;
+  margin-right:10px;
+  margin-bottom:29px;
+`;
+
+const TextArea = styled.textarea`
+  width: 100%;
+  height: auto;
+  min-height: 400px; 
+  border: none;
+  padding: 5px;
+  resize: none; 
+  font-size:13px;
+  line-height: 1.5; 
+
+  &:focus {
+    outline: none;    
+    border: none;  
+    box-shadow: none;
+  }
+
+  @media (max-width: 554px) {
+    min-height: 50px;
+  }
+`;
 
 const getCurrentDate = () => {
   const today = new Date();
   const year = today.getFullYear();
-  const month = String(today.getMonth() + 1).padStart(2, "0");
+  const month = String(today.getMonth() + 1).padStart(2, "0"); 
   const day = String(today.getDate()).padStart(2, "0");
   return `${year}-${month}-${day}`;
 };
 
+
 function WriteReview() {
-  const [ratings, setRatings] = useState([false, false, false, false, false]);
-  const [previews, setPreviews] = useState([]);
+  const [ratings, setRatings] = useState([false, false, false, false, false]); // 별점
+  const [previews, setPreviews] = useState([]); // 첨부하는 이미지 미리보기
+  const [selectPet, setSelectPet] = useState(""); // 펫 선택
+  const [userNickname, setUserNickname] = useState('내가 진짜'); //나중에 zustand로 받아야 와하는 유저 닉네임
+  const [userImage, setUserImage] = useState(''); // 예시 이미지 URL, 나중에 zustand로 받아야 와하는 유저 이미지
+  const [selectKeywords, setSelectKeywords] = useState([]); // 키워드 선택
+  const [content, setContent] = useState(""); // 리뷰 내용
+  const [visitedAt, setVisitedAt] = useState(getCurrentDate()); // 방문한 날짜
 
   const handleImageUpload = (e) => {
     const files = Array.from(e.target.files);
@@ -244,11 +317,57 @@ function WriteReview() {
     setPreviews((prevPreviews) => prevPreviews.filter((_, i) => i !== index));
   };
 
+  const [text, setText] = useState('');
+  const maxLength = 500;
+
+  const handleChange = (e) => {
+    if (e.target.value.length > maxLength) {
+      AlertDialog({
+        mode: "alert",
+        title: "선택 제한",
+        text: `최대 ${maxLength}자까지만 작성 가능합니다.`,
+        confirmText: "닫기" 
+      });
+    } else {
+      setText(e.target.value);  
+    }
+  };
+
+  const handleSubmit = async () => {
+    const placeId = 1; // 예시로 1번 장소
+    const score = ratings.filter((rating) => rating).length; // 별점 
+    const media = previews;
+    const keywords = selectKeywords; // 선택한 키워드들 > 공통 코드로 설정해야함
+    const pets = [1, 2, 3]; // 예시로 넣음
+
+    const reviewData = {
+      placeId,
+      content,
+      score,
+      media,
+      keywords,
+      pets,
+      visitedAt,
+    };
+
+    try {
+      const response = await axios.post("http://localhost:8080/api/v1/review", reviewData, {
+        headers: {
+          "Content-Type": "application/json",
+          "Authorization": `Bearer ${authToken}`,
+        },
+      });
+      console.log("리뷰 등록 성공:", response.data);
+    } catch (error) {
+      console.error("리뷰 등록 실패:", error);
+    }
+  };
+
   return (
     <WriteReviewAllContainer>
       <WriteReviewContainer>
         <PlaceTitle>가평 트리하우스</PlaceTitle>
-        <WriteReviewDate>2024-06-20</WriteReviewDate>
+        <WriteReviewDate>{getCurrentDate()}</WriteReviewDate>
       </WriteReviewContainer>
       <SelectPlaceOptionContainer>
         <WhatPointLike>어떤 점이 좋았나요?</WhatPointLike>
@@ -256,12 +375,16 @@ function WriteReview() {
         <SelectWarning>*이 장소에 맞는 키워드를 골라주세요 (1개~3개)</SelectWarning>
         <PreferenceFavoriteOptionList />
         <UserInfoContainer>
-          <UserImg />
-          <UserNickname>내가 진짜</UserNickname>
+          <UserImg/> 
+          <UserNickname>{userNickname || "내가 진짜"}</UserNickname>
         </UserInfoContainer>
         <UserQuestionContainer>
           <Question>함께한 댕댕이를 선택해주세요</Question>
-          <PetSelection />
+          <PetSelection value={selectPet} onChange={(e) => setSelectPet(e.target.value)}>
+          <option value="pet1">댕댕이 1</option>
+          <option value="pet2">댕댕이 2</option>
+          <option value="pet3">댕댕이 3</option>
+        </PetSelection>
         </UserQuestionContainer>
         <UserQuestionContainer>
           <Question>방문한 날짜를 선택해주세요</Question>
@@ -305,7 +428,14 @@ function WriteReview() {
             />
           </AddImg>
         </AddImgContainer>
-        <TextContainer />
+        <TextDescriptionContainer>
+      <QuestionBox>리뷰를 작성해주세요</QuestionBox>
+      <CountText>{text.length}자 | 최대 500자</CountText>
+    </TextDescriptionContainer>
+    <DivisionLine />
+    <TextArea type='text' placeholder='경험을 공유해주세요' value={text} onChange={handleChange}/>
+    <DivisionLine />
+    <ConfirmBtn onClick={handleSubmit} marginBottom="29px" label="작성 완료" />
       </SelectPlaceOptionContainer>
     </WriteReviewAllContainer>
   );
